:recycle: :truck: split models into multiple files
