+++ web/app/src/lib/context/profile.ts
@@ -1,5 +1,5 @@
 import { createContext, useContext } from 'react'
-import { ProfileModel } from '@/lib/models'
+import { ProfileModel } from '@/lib/models/auth'
 
 const ProfileContext = createContext<ProfileModel>(null!)
