+++ web/app/src/lib/models/forwarder/http.ts
@@ -1,0 +1,5 @@
+export type HttpForwarderModel = {
+  type: 'http'
+  url: string
+  headers: Record<string, string>
+}
