+++ web/app/src/lib/models/forwarder/index.ts
@@ -1,0 +1,14 @@
+import { HttpForwarderModel } from '@/lib/models/forwarder/http'
+
+export type ForwarderModel = {
+  config: ForwarderConfigModel
+}
+
+export const ForwarderTypeValues = [
+  { key: 'http', label: 'Webhook' },
+] as const
+
+export type ForwarderConfigModel =
+  | HttpForwarderModel
+
+export type ForwarderTypes = ForwarderConfigModel['type']
