+++ web/app/src/lib/models/log.ts
@@ -1,0 +1,4 @@
+export type LogEntryModel = {
+  timestamp: Date
+  fields: Record<string, string>
+}
