+++ web/app/src/lib/models/pipeline.ts
@@ -1,0 +1,6 @@
+import { ReactFlowJsonObject } from '@xyflow/react'
+
+export type PipelineModel = {
+  nodes: ReactFlowJsonObject['nodes'],
+  edges: ReactFlowJsonObject['edges'],
+}
