+++ web/app/src/lib/models/storage.ts
@@ -1,0 +1,5 @@
+export type StreamConfigModel = {
+  indexed_fields: string[]
+  ttl: number
+  size: number
+}
