+++ web/app/src/views/app/admin/role-list/create-btn/index.tsx
@@ -6,7 +6,7 @@ import AddIcon from '@mui/icons-material/Add'
 
 import Button from '@mui/material/Button'
 
-import { RoleModel } from '@/lib/models'
+import { RoleModel } from '@/lib/models/auth'
 
 import { RoleFormModal } from './modal'
